Add kv "entry" actions
An "entry" is a JSON object with properties. Multiple entries are stored
under a JSON serialized object (as created by st2.kv.set_object) in the
datastore.

You can upsert or delete properties within an entry. And, you can define a
fallback entry (which is named "default" by default) to allow getting
properties from an entry, and using the fallback entry to provide defaults for
any properties that are undfined in the entry.

One use-case for objects like this is to store environment-specific (eg dev,
stage, prod, etc) variables with global defaults. Each environment would be
an entry and "global" could be the fallback entry. The variables would be
properties inside the entries.

Signed-off-by: Jacob Floyd <cognifloyd@gmail.com>